Made all functions return double

+++ src/Math.h
@@ -7,22 +7,22 @@ namespace Nash
         namespace Arithmetic
         {
             // Returns __num1 + __num2
-            int Add(int __num1, int __num2)
+            double Add(double __num1, double __num2)
             {
                 return __num1 + __num2;
             }
 
-            int Subtract(int __num1, int __num2)
+            double Subtract(double __num1, double __num2)
             {
                 return __num1 - __num2;
             }
 
-            int Multiply(int __num1, int __num2)
+            double Multiply(double __num1, double __num2)
             {
                 return __num1 * __num2;
             }
 
-            int Divide(int __num1, int __num2)
+            double Divide(double __num1, double __num2)
             {
                 return __num1 / __num2;
             }
@@ -30,59 +30,59 @@ namespace Nash
         namespace Trigonometry
         {
             // Returns the sine of the number of radians entered
-            int Sine(int __radians)
+            double Sine(double __radians)
             {
-                return (int)sin((double)__radians);
+                return sin(__radians);
             }
             // Returns the cosine of the number of radians entered
-            int Cosine(int __radians)
+            double Cosine(double __radians)
             {
-                return (int)cos((double)__radians);
+                return cos(__radians);
             }
             // Returns the tangent of the number of radians entered
-            int Tangent(int __radians)
+            double Tangent(double __radians)
             {
-                return (int)tan((double)__radians);
+                return tan(__radians);
             }
             // Returns the arc sine of the number of radians entered
-            int ArcSine(int __radians)
+            double ArcSine(double __radians)
             {
-                return (int)asin((double)__radians);
+                return asin(__radians);
             }
             // Returns the arc cosine of the number of radians entered
-            int ArcCosine(int __radians)
+            double ArcCosine(double __radians)
             {
-                return (int)acos((double)__radians);
+                return acos(__radians);
             }
             // Returns the arc tangent of the number of radians entered
-            int ArcTangent(int __radians)
+            double ArcTangent(double __radians)
             {
-                return (int)atan((double)__radians);
+                return atan(__radians);
             }
             // Returns the arc tangent of the number of y/x
-            int YXArcTangent(int __x, int __y)
+            double YXArcTangent(double __x, double __y)
             {
-                return (int)atan2((double)__y,(double) __x);
+                return atan2(__y, __x);
             }
         }
         namespace BasicMath
         {
             // Return the square root of the number input
-            int root(int __num)
+            double root(double __num)
             {
-                return (int)sqrt((double)__num);
+                return sqrt(__num);
             }
 
             // Return the cube root of the number input
-            int cube_root(int __num)
+            double cube_root(double __num)
             {
-                return (int)cbrt((double)__num);
+                return cbrt(__num);
             }
 
             // Return the input ^ n
-            int to_power_of_n(int __num, int __n)
+            double to_power_of_n(double __num, double __n)
             {
-                int result = __num * __num;
+                double result = __num * __num;
                 for (int i = 0; i < __num - 1; i++)
                 {
                     result = result * __num;   
@@ -91,20 +91,20 @@ namespace Nash
             }
 
             // Returns the rounded version of the number input
-            int round_num(int __num)
+            double round_num(double __num)
             {
-                return (int)round((double)__num);
+                return round(__num);
             }
             // Returns natural logarithm of the number input (assumes that base is e if not supplied)
-            int logarithm(int __num, int __base = NULL)
+            double logarithm(double __num, double __base = NULL)
             {
-                if (__base == NULL)    { return (int)log((double)__num); }
+                if (__base == NULL)    { return log(__num); }
                 else                   { return (__num > __base - 1) ? 1 + logarithm(__num / __base, __base) : 0; }
             }
             // Returns absolute value of input (usage => value.absolute() )
-            int absolute(int __num)
+            double absolute(double __num)
             {
-                return (int)abs((double)__num);
+                return abs(__num);
             }
 
         }
